requirements.txt
pandas
streamlit
scikit-learn
numpy
google.cloud
vertexai